LRCI-1763 Move the segment splitting logic into BatchTestClassGroup
